refactor(chat-ui): clean KaTex utils
