+++ frontend/src/routes/chat/+page.svelte
@@ -1,10 +1,9 @@
 <script lang="ts">
     import { SearchIcon, ChevronDownIcon, CheckIcon } from "svelte-feather-icons";
     import { onDestroy } from 'svelte';
-    import { marked } from 'marked';
-    import katex from "katex";
     import "katex/dist/katex.min.css";
     import { processReasoningMessage, processOutputMessage, toggleReasoning } from '$lib/utils/chat';
+    import { renderLatex, renderContent } from '$lib/utils/katex';
 
     let userQuery = '';
     let conversationId: string | null = null;
@@ -88,53 +87,6 @@
             console.error('SSE error:', err);
             eventSource?.close();
         });
-    }
-
-    // Function to render LaTeX blocks
-    function renderLatex(content: string) {
-        
-        const tempDiv = document.createElement('div');
-        tempDiv.innerHTML = content;
-
-        // Find all LaTeX blocks and render them
-        const mathElements = tempDiv.getElementsByClassName('math');
-        Array.from(mathElements).forEach(element => {
-            const latex = element.textContent || '';
-            try {
-                const isDisplay = element.classList.contains('math-display');
-                const rendered = katex.renderToString(latex, {
-                    displayMode: isDisplay,
-                    throwOnError: false,
-                    strict: false
-                });
-                element.innerHTML = rendered;
-            } catch (error) {
-                console.error('LaTeX rendering error:', error);
-            }
-        });
-
-        return tempDiv.innerHTML;
-    }
-  
-    // Function to render Markdown with LaTeX support
-    function renderMarkdown(content: string) {
-
-        // Handle display math: \[ ... \] before markdown processing
-        content = content.replace(/\\\[([\s\S]*?)\\\]/g, (match, latex) => {
-            return `<div class="math math-display">${latex}</div>`;
-        });
-
-        // Handle inline math: \( ... \) before markdown processing
-        content = content.replace(/\\\(([\s\S]*?)\\\)/g, (match, latex) => {
-            return `<span class="math math-inline">${latex}</span>`;
-        });
-
-        return marked(content);
-    }
-
-    function renderContent(text: string) {
-        const markdownRendered = renderMarkdown(text) as string;
-        return renderLatex(markdownRendered);
     }
 
     onDestroy(() => {
